pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-ai-inference~=1.0.0b7",
    "azure-ai-projects>=1.0.0",
    "azure-identity~=1.20.0",
    "dotenv>=0.9.9",
]
